api finished

+++ main.py
@@ -3,6 +3,8 @@ from threading import Thread
 from flask_restful import Resource, Api
 import xlrd
 import requests
+import os
+
 
 app = Flask(__name__)
 api = Api(app)
@@ -29,9 +31,26 @@ class Info(Resource):
         self.fileName = self.request_salaries()
 
     def get(self, school, lastName, job, fte):
-        return {
-            "Hello": "World"
-        }
+        # School Name: column 2
+        # FTE: column 3
+        # Annual Salary and benefits: column 5 + 7
+        # Job Title: column 9
+        # Name: column 10
+
+        wb = xlrd.open_workbook(self.fileName)
+        sheet = wb.sheet_by_index(0)
+
+        matches = []
+        for row_idx in range(1, sheet.nrows):
+            row = sheet.row_values(row_idx)
+
+            if (school == 'Default' or school.lower() in row[2].lower()) and \
+            (fte == 'Default' or row[3] == float(fte)) and \
+            (job == 'Default' or job.lower() in row[9].lower()) and \
+            (lastName == 'Default' or lastName in row[10].lower()):
+                matches.append((row[2], row[3], row[5], row[7], row[9], row[10]))
+
+        return jsonify(matches)
 
     def request_salaries(self):
         url = "https://www.cps.edu/about/finance/employee-position-files/"
@@ -43,16 +62,17 @@ class Info(Resource):
         fileName = response[filePos + 65 : filePos + 100]
 
         #download file
-        r = requests.get(fileURL)
-        with open(fileName, 'wb') as f:
-            f.write(r.content)
+        if not os.path.isfile(fileName):
+            r = requests.get(fileURL)
+            with open(fileName, 'wb') as f:
+                f.write(r.content)
 
         return fileName
 
 
 #api endpoint
 api.add_resource(Empty, '/')
-api.add_resource(Info, '/<string:school>/<string:lastName>/<string:job>/<float:fte>')
+api.add_resource(Info, '/<string:school>/<string:lastName>/<string:job>/<string:fte>')
 
 
 if __name__ == '__main__':
